Fixed Delete methods on Commission Sales Employee, Hourly Employee, and Salary Employee

+++ project.sql
@@ -57,8 +57,8 @@ CREATE TABLE if not exists HourlyEmployee(
     `lastName` VARCHAR(200) NOT NULL ,
     `position` VARCHAR(200) NOT NULL ,
     `department` VARCHAR(200) NOT NULL ,
-    `hoursPerWeek` DECIMAL(3,2) NOT NULL,
-    `hourlyRate` DECIMAL(3,2) NOT NULL,
+    `hoursPerWeek` DECIMAL(4,2) NOT NULL,
+    `hourlyRate` DECIMAL(9,2) NOT NULL,
     PRIMARY KEY(id)
 ) ENGINE=INNODB;
 
@@ -68,7 +68,7 @@ CREATE TABLE if not exists SalaryEmployee(
     `lastName` VARCHAR(200) NOT NULL ,
     `position` VARCHAR(200) NOT NULL ,
     `department` VARCHAR(200) NOT NULL ,
-    `salary` DOUBLE(3,2) NOT NULL,
+    `salary` DECIMAL(15,2) NOT NULL,
     PRIMARY KEY(id)
 ) ENGINE=INNODB;
